+++ src/posteriorpln.c
@@ -139,7 +139,7 @@ void gpln (int *pop,
       lpm[i]=i*gammart+lgamma(ni+i+1.)-lgamma(i+1.);
 //    Add in the (log) prior on m: P(m)
       lpm[i]=lpm[i]+lpriorm[i];
-      if(lpm[i] > tU) tU = lpm[i];
+      if(lpm[i] > tU) tU = (int)lpm[i];
     }
     for (i=0; i<imaxm; i++){
       lpm[i]=exp(lpm[i]-tU);
